[FEAT]: 숏폼 업로드 Reader/Writer 분리, 상태 추가
